Core: Add raw data recording mode
