+++ BililiveRecorder.Core/Config/V2/RecordMode.cs
@@ -1,0 +1,8 @@
+namespace BililiveRecorder.Core.Config.V2
+{
+    public enum RecordMode : int
+    {
+        Standard = 0,
+        RawData = 1,
+    }
+}
