+++ BililiveRecorder.Core/Recording/RecordTaskFactory.cs
@@ -1,4 +1,5 @@
 using System;
+using BililiveRecorder.Core.Config.V2;
 using Microsoft.Extensions.DependencyInjection;
 
 namespace BililiveRecorder.Core.Recording
@@ -13,6 +14,10 @@ namespace BililiveRecorder.Core.Recording
         }
 
         public IRecordTask CreateRecordTask(IRoom room) =>
-            ActivatorUtilities.CreateInstance<RecordTask>(this.serviceProvider, room);
+            room.RoomConfig.RecordMode switch
+            {
+                RecordMode.RawData => ActivatorUtilities.CreateInstance<RawDataRecordTask>(this.serviceProvider, room),
+                _ => ActivatorUtilities.CreateInstance<StandardRecordTask>(this.serviceProvider, room)
+            };
     }
 }
