+++ BililiveRecorder.WPF/Pages/SettingsPage.xaml
@@ -41,6 +41,18 @@
                     </StackPanel>
                 </StackPanel>
             </GroupBox>
+            <GroupBox Header="录制模式">
+                <StackPanel>
+                    <StackPanel>
+                        <RadioButton GroupName="RecordMode" Content="标准模式"
+                                     IsChecked="{Binding Path=RecordMode, Converter={StaticResource EnumToBooleanConverter},
+                            ConverterParameter={x:Static confiv2:RecordMode.Standard}}" />
+                        <RadioButton GroupName="RecordMode" Content="原始数据模式"
+                                     IsChecked="{Binding Path=RecordMode, Converter={StaticResource EnumToBooleanConverter},
+                            ConverterParameter={x:Static confiv2:RecordMode.RawData}}" />
+                    </StackPanel>
+                </StackPanel>
+            </GroupBox>
             <GroupBox Header="{l:Loc Settings_Splitting_Title}">
                 <StackPanel>
                     <RadioButton GroupName="Splitting" Name="CutDisabledRadioButton" Content="{l:Loc Settings_Splitting_RadioButton_Disabled}"
